incorporate sidebar into the privateNav, and create PrivateLayout
