+++ client/components/PrivateNav.jsx
@@ -1,14 +1,22 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
+import { RiPictureInPictureExitLine } from "react-icons/ri";
 import { HiOutlineMenu } from "react-icons/hi";
 
-import Sidebar from "../components/Sidebar";
 import logo from "../../docs/assets/images/logo.png";
 
 const PrivateNav = (props) => {
+  const [menu, setMenu] = useState("");
   const [activeSideBar, setActiveSideBar] = useState(false);
+
+  // const navigate = useNavigate();
+  const handleLogout = () => {
+    setMenu("Logout");
+    // process logout in backend
+  };
+
   return (
-    <main className="h-screen justify-center items-center relative bg-priwhite">
+    <main className="justify-center items-center relative bg-priwhite">
       <div className="w-full h-14 flex flex-row items-center pt-10 pb-10 left-6 border-b border-thdwhite">
         <HiOutlineMenu
           className="w-6 ml-6 mr-6 font-bold size-14"
@@ -20,13 +28,87 @@ const PrivateNav = (props) => {
           Hi, {props.user.firstName}
         </span>
       </div>
+      
       <div>
         {activeSideBar ? (
-          <div className="w-72 fixed bg-priwhite">
-            <Sidebar user={props.user} />
-          </div>
+          <>
+            <div className="h-screen w-72 bg-priblue text-priwhite fixed flex-col top-0 left-0 m-0 flex shadow-lg">
+              <RiPictureInPictureExitLine
+                className="size-6 mt-4 ml-60 text-thdwhite"
+                onClick={() => setActiveSideBar(false)}
+              />
+              <div className="flex flex-col justify-between items-center mt-40">
+                <Link to="/profile" className="items-center">
+                  <span
+                    className={`flex items-center justify-center top-4 font-bold rounded-full h-16 w-16 bg-thdwhite text-priblue ${
+                      props.user.firstName.length <= 6
+                        ? "text-xl"
+                        : "h-16 w-16 text-sm"
+                    }`}
+                  >
+                    {props.user.firstName}
+                  </span>
+                </Link>
+              </div>
+
+              <div>
+                <ul>
+                  <li
+                    className="flex flex-col mt-20 ml-20"
+                    onClick={() => setMenu("Profile")}
+                  >
+                    <Link style={{ textDecoration: "none" }} to={"/profile"}>
+                      Profile
+                      {menu === "Profile" ? <hr className="w-12" /> : <></>}
+                    </Link>
+                  </li>
+                  <li
+                    className="flex flex-col mt-10 ml-20"
+                    onClick={() => setMenu("Project")}
+                  >
+                    <Link
+                      style={{ textDecoration: "none" }}
+                      to={"/profile/projectlist"}
+                    >
+                      Project
+                      {menu === "Project" ? <hr className="w-14" /> : <></>}
+                    </Link>
+                  </li>
+                  <li
+                    className="flex flex-col mt-10 ml-20"
+                    onClick={() => setMenu("Payment")}
+                  >
+                    <Link style={{ textDecoration: "none" }} to={"/profile"}>
+                      Payment
+                      {menu === "Payment" ? <hr className="w-16" /> : <></>}
+                    </Link>
+                  </li>
+                  <li
+                    className="flex flex-col mt-10 ml-20"
+                    onClick={() => setMenu("Message")}
+                  >
+                    <Link style={{ textDecoration: "none" }} to={"/profile"}>
+                      Message
+                      {menu === "Message" ? <hr className="w-16" /> : <></>}
+                    </Link>
+                  </li>
+                </ul>
+              </div>
+              <div>
+                <Link
+                  to="/profile"
+                  onClick={handleLogout}
+                  className="items-center"
+                >
+                  <span className="flex flex-col mt-10 ml-20">
+                    Logout{menu === "Logout" ? <hr className="w-12" /> : <></>}
+                  </span>
+                </Link>
+              </div>
+            </div>
+          </>
         ) : (
-          <div className="w-0"></div>
+          <div></div>
         )}
       </div>
     </main>
