+++ client/subpage/Profile.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { Container, Box } from "@mui/material";
 
-import PrivateNav from "../components/PrivateNav";
+import PrivateLayout from "../components/PrivateLayout";
 import axios from "../hook/axios";
 
 const Profile = () => {
@@ -31,44 +31,39 @@ const Profile = () => {
 
   return (
     <div>
-      <PrivateNav user={currUser}/>
-      <Container
-      // sx={{
-      //   display: "flex",
-      //   flexDirection: "row",
-      //   justifyContent: "space-between",
-      // }}
-      >
-        <Box
-          height={100}
-          width="100%"
-          borderRadius={0.5}
-          my={4}
-          display={"flex"}
-          alignItems="center"
-          flexDirection={"row"}
-          justifyContent={"center"}
-          gap={4}
-          p={2}
-          sx={{ marginLeft: "48px" }}
-          className="text-priblue "
-        >
-          <span className="text-xl"> Current Balance: </span>
-          <span className="font-bold text-4xl">$3,000</span>
-        </Box>
+      <PrivateLayout>
+        <Container>
+          <Box
+            height={100}
+            width="100%"
+            borderRadius={0.5}
+            my={4}
+            display={"flex"}
+            alignItems="center"
+            flexDirection={"row"}
+            justifyContent={"center"}
+            gap={4}
+            p={2}
+            sx={{ marginLeft: "48px" }}
+            className="text-priblue "
+          >
+            <span className="text-xl"> Current Balance: </span>
+            <span className="font-bold text-4xl">$3,000</span>
+          </Box>
 
-        <Box
-          height={300}
-          width={1200}
-          my={4}
-          alignItems="center"
-          gap={4}
-          p={2}
-          sx={{ border: "2px solid grey", marginRight: "48px" }}
-        >
-          Project Overview
-        </Box>
-      </Container>
+          <Box
+            height={300}
+            width={1200}
+            my={4}
+            alignItems="center"
+            gap={4}
+            p={2}
+            sx={{ border: "2px solid grey", marginRight: "48px" }}
+          >
+            Project Overview
+          </Box>
+        </Container>
+      </PrivateLayout>
     </div>
   );
 };
